Split semantic bundle task tests
